move to resource parsing exception
this move the exception to resource parsing time.

since cookbooks that depend on compat_resource are written in Chef 12.5
style and all cookbooks that are synched have all their resources
loaded, i think this won't fix anyone.  the major benefit of throwing
the exception here is that users more clearly see which code is
inherently not supportable on Chef 11 and might stop blaming the
messenger.