fix recursive resource cloning on medium-age chef-clients
since we patched the outer run context to do recursive lookups,
we also need to upgrade the resource_builder for all resources
in order to have it call lookup_local instead of doing a
lookup_recursive.

need to figure out if we need to do anything else for pre-
resource_builder chef clients as well.