back to a deprecation warning
i think this is actually better.  users on old chef + chef-dk omnibus
binaries will be broken by slurping down the new compat_resource
cookbook.  they only path out of it is upgrading their chef + chef-dk
packages -- but the whole point of compat_resource is so that people
who don't ever upgrade don't block community cookbook development.

so we need to make this work.  we're already monkeypatching over the
top of a dozen old chef-client versions, we might as well monkeypatch
over the top of our old monkeypatches as well.

the general rule we're going to start following around here is that if
you're making the choice to monkeypatch code you need to assert that
you always win any argument.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>